move erigon code to a single file
