+++ docker/test_env/erigon.go
@@ -22,12 +22,14 @@ import (
 	"github.com/smartcontractkit/chainlink-testing-framework/docker"
 	"github.com/smartcontractkit/chainlink-testing-framework/logging"
 	"github.com/smartcontractkit/chainlink-testing-framework/mirror"
+	"github.com/smartcontractkit/chainlink-testing-framework/utils/templates"
 	"github.com/smartcontractkit/chainlink-testing-framework/utils/testcontext"
 )
 
 const defaultErigonPosImage = "thorax/erigon:v2.56.2"
+const defaultErigonPowImage = "thorax/erigon:v2.40.0"
 
-type ErigonPos struct {
+type Erigon struct {
 	EnvComponent
 	ExternalHttpUrl      string
 	InternalHttpUrl      string
@@ -42,9 +44,9 @@ type ErigonPos struct {
 	t                    *testing.T
 }
 
-func NewErigonPos(networks []string, chainConfg *EthereumChainConfig, generatedDataHostDir string, consensusLayer ConsensusLayer, opts ...EnvComponentOption) (*ErigonPos, error) {
+func NewErigonPos(networks []string, chainConfg *EthereumChainConfig, generatedDataHostDir string, consensusLayer ConsensusLayer, opts ...EnvComponentOption) (*Erigon, error) {
 	parts := strings.Split(defaultErigonPosImage, ":")
-	g := &ErigonPos{
+	g := &Erigon{
 		EnvComponent: EnvComponent{
 			ContainerName:    fmt.Sprintf("%s-%s", "erigon-pos", uuid.NewString()[0:8]),
 			Networks:         networks,
@@ -65,14 +67,41 @@ func NewErigonPos(networks []string, chainConfg *EthereumChainConfig, generatedD
 	return g, nil
 }
 
-func (g *ErigonPos) WithTestInstance(t *testing.T) ExecutionClient {
+func NewErigonPow(networks []string, chainConfg *EthereumChainConfig, opts ...EnvComponentOption) (*Erigon, error) {
+	parts := strings.Split(defaultErigonPowImage, ":")
+	g := &Erigon{
+		EnvComponent: EnvComponent{
+			ContainerName:    fmt.Sprintf("%s-%s", "erigon-pow", uuid.NewString()[0:8]),
+			Networks:         networks,
+			ContainerImage:   parts[0],
+			ContainerVersion: parts[1],
+		},
+		chainConfg: chainConfg,
+		l:          logging.GetTestLogger(nil),
+	}
+	g.SetDefaultHooks()
+	for _, opt := range opts {
+		opt(&g.EnvComponent)
+	}
+	// if the internal docker repo is set then add it to the version
+	g.EnvComponent.ContainerImage = mirror.AddMirrorToImageIfSet(g.EnvComponent.ContainerImage)
+	return g, nil
+}
+
+func (g *Erigon) WithTestInstance(t *testing.T) ExecutionClient {
 	g.l = logging.GetTestLogger(t)
 	g.t = t
 	return g
 }
 
-func (g *ErigonPos) StartContainer() (blockchain.EVMNetwork, error) {
-	r, err := g.getContainerRequest()
+func (g *Erigon) StartContainer() (blockchain.EVMNetwork, error) {
+	var r *tc.ContainerRequest
+	var err error
+	if g.consensusLayer != "" {
+		r, err = g.getPosContainerRequest()
+	} else {
+		r, err = g.getPowContainerRequest()
+	}
 	if err != nil {
 		return blockchain.EVMNetwork{}, err
 	}
@@ -99,9 +128,14 @@ func (g *ErigonPos) StartContainer() (blockchain.EVMNetwork, error) {
 	if err != nil {
 		return blockchain.EVMNetwork{}, err
 	}
-	executionPort, err := ct.MappedPort(testcontext.Get(g.t), NatPort(ETH2_EXECUTION_PORT))
-	if err != nil {
-		return blockchain.EVMNetwork{}, err
+
+	if g.consensusLayer != "" {
+		executionPort, err := ct.MappedPort(testcontext.Get(g.t), NatPort(ETH2_EXECUTION_PORT))
+		if err != nil {
+			return blockchain.EVMNetwork{}, err
+		}
+		g.InternalExecutionURL = FormatHttpUrl(g.ContainerName, ETH2_EXECUTION_PORT)
+		g.ExternalExecutionURL = FormatHttpUrl(host, executionPort.Port())
 	}
 
 	g.Container = ct
@@ -109,11 +143,13 @@ func (g *ErigonPos) StartContainer() (blockchain.EVMNetwork, error) {
 	g.InternalHttpUrl = FormatHttpUrl(g.ContainerName, TX_GETH_HTTP_PORT)
 	g.ExternalWsUrl = FormatWsUrl(host, httpPort.Port())
 	g.InternalWsUrl = FormatWsUrl(g.ContainerName, TX_GETH_HTTP_PORT)
-	g.InternalExecutionURL = FormatHttpUrl(g.ContainerName, ETH2_EXECUTION_PORT)
-	g.ExternalExecutionURL = FormatHttpUrl(host, executionPort.Port())
 
 	networkConfig := blockchain.SimulatedEVMNetwork
-	networkConfig.Name = fmt.Sprintf("Simulated Ethereum-PoS (erigon + %s)", g.consensusLayer)
+	if g.consensusLayer != "" {
+		networkConfig.Name = fmt.Sprintf("Simulated Ethereum-PoS (erigon + %s)", g.consensusLayer)
+	} else {
+		networkConfig.Name = "Simulated Ethereum-PoW (erigon)"
+	}
 	networkConfig.URLs = []string{g.ExternalWsUrl}
 	networkConfig.HTTPURLs = []string{g.ExternalHttpUrl}
 
@@ -123,45 +159,45 @@ func (g *ErigonPos) StartContainer() (blockchain.EVMNetwork, error) {
 	return networkConfig, nil
 }
 
-func (g *ErigonPos) GetInternalExecutionURL() string {
+func (g *Erigon) GetInternalExecutionURL() string {
 	return g.InternalExecutionURL
 }
 
-func (g *ErigonPos) GetExternalExecutionURL() string {
+func (g *Erigon) GetExternalExecutionURL() string {
 	return g.ExternalExecutionURL
 }
 
-func (g *ErigonPos) GetInternalHttpUrl() string {
+func (g *Erigon) GetInternalHttpUrl() string {
 	return g.InternalHttpUrl
 }
 
-func (g *ErigonPos) GetInternalWsUrl() string {
+func (g *Erigon) GetInternalWsUrl() string {
 	return g.InternalWsUrl
 }
 
-func (g *ErigonPos) GetExternalHttpUrl() string {
+func (g *Erigon) GetExternalHttpUrl() string {
 	return g.ExternalHttpUrl
 }
 
-func (g *ErigonPos) GetExternalWsUrl() string {
+func (g *Erigon) GetExternalWsUrl() string {
 	return g.ExternalWsUrl
 }
 
-func (g *ErigonPos) GetContainerName() string {
+func (g *Erigon) GetContainerName() string {
 	return g.ContainerName
 }
 
-func (g *ErigonPos) GetContainer() *tc.Container {
+func (g *Erigon) GetContainer() *tc.Container {
 	return &g.Container
 }
 
-func (g *ErigonPos) getContainerRequest() (*tc.ContainerRequest, error) {
+func (g *Erigon) getPosContainerRequest() (*tc.ContainerRequest, error) {
 	initFile, err := os.CreateTemp("", "init.sh")
 	if err != nil {
 		return nil, err
 	}
 
-	initScriptContent, err := g.buildInitScript()
+	initScriptContent, err := g.buildPosInitScript()
 	if err != nil {
 		return nil, err
 	}
@@ -211,30 +247,141 @@ func (g *ErigonPos) getContainerRequest() (*tc.ContainerRequest, error) {
 	}, nil
 }
 
-func (g *ErigonPos) WaitUntilChainIsReady(ctx context.Context, waitTime time.Duration) error {
+func (g *Erigon) getPowContainerRequest() (*tc.ContainerRequest, error) {
+	keystoreDir, err := os.MkdirTemp("", "keystore")
+	if err != nil {
+		return nil, err
+	}
+
+	generatedData, err := generateKeystoreAndExtraData(keystoreDir)
+	if err != nil {
+		return nil, err
+	}
+
+	genesisJsonStr, err := templates.GenesisJsonTemplate{
+		ChainId:     fmt.Sprintf("%d", g.chainConfg.ChainID),
+		AccountAddr: []string{RootFundingAddr, generatedData.minerAccount.Address.Hex()},
+		Consensus:   templates.GethGenesisConsensus_Ethash,
+	}.String()
+	if err != nil {
+		return nil, err
+	}
+
+	initFile, err := os.CreateTemp("", "init.sh")
+	if err != nil {
+		return nil, err
+	}
+
+	initScriptContent, err := g.buildPowInitScript(generatedData.minerAccount.Address.Hex())
+	if err != nil {
+		return nil, err
+	}
+
+	_, err = initFile.WriteString(initScriptContent)
+	if err != nil {
+		return nil, err
+	}
+
+	genesisFile, err := os.CreateTemp("", "genesis_json")
+	if err != nil {
+		return nil, err
+	}
+	_, err = genesisFile.WriteString(genesisJsonStr)
+	if err != nil {
+		return nil, err
+	}
+	key1File, err := os.CreateTemp(keystoreDir, "key1")
+	if err != nil {
+		return nil, err
+	}
+	_, err = key1File.WriteString(RootFundingWallet)
+	if err != nil {
+		return nil, err
+	}
+
+	return &tc.ContainerRequest{
+		Name:          g.ContainerName,
+		Image:         g.GetImageWithVersion(),
+		Networks:      g.Networks,
+		ImagePlatform: "linux/x86_64",
+		ExposedPorts:  []string{NatPortFormat(TX_GETH_HTTP_PORT)},
+		WaitingFor: tcwait.ForAll(
+			tcwait.ForLog("Started P2P networking").
+				WithStartupTimeout(120*time.Second).
+				WithPollInterval(1*time.Second),
+			NewWebSocketStrategy(NatPort(TX_GETH_HTTP_PORT), g.l),
+		),
+		User: "0:0",
+		Entrypoint: []string{
+			"sh",
+			"/home/erigon/init.sh",
+		},
+		Files: []tc.ContainerFile{
+			{
+				HostFilePath:      initFile.Name(),
+				ContainerFilePath: "/root/init.sh",
+				FileMode:          0644,
+			},
+			{
+				HostFilePath:      genesisFile.Name(),
+				ContainerFilePath: "/root/genesis.json",
+				FileMode:          0644,
+			},
+			{
+				HostFilePath:      initFile.Name(),
+				ContainerFilePath: "/home/erigon/init.sh",
+				FileMode:          0744,
+			},
+		},
+		HostConfigModifier: func(hostConfig *container.HostConfig) {
+			hostConfig.Mounts = append(hostConfig.Mounts, mount.Mount{
+				Type:     mount.TypeBind,
+				Source:   keystoreDir,
+				Target:   "/root/.local/share/erigon/keystore/",
+				ReadOnly: false,
+			},
+			)
+		},
+		LifecycleHooks: []tc.ContainerLifecycleHooks{
+			{
+				PostStarts: g.PostStartsHooks,
+				PostStops:  g.PostStopsHooks,
+			},
+		},
+	}, nil
+}
+
+func (g *Erigon) WaitUntilChainIsReady(ctx context.Context, waitTime time.Duration) error {
+	if g.consensusLayer == "" {
+		return nil
+	}
 	waitForFirstBlock := tcwait.NewLogStrategy("Built block").WithPollInterval(1 * time.Second).WithStartupTimeout(waitTime)
 	return waitForFirstBlock.WaitUntilReady(ctx, *g.GetContainer())
 }
 
-func (c *ErigonPos) getExtraExecutionFlags() (string, error) {
-	version, err := GetComparableVersionFromDockerImage(c.GetImageWithVersion())
+func (g *Erigon) getExtraExecutionFlags() (string, error) {
+	version, err := GetComparableVersionFromDockerImage(g.GetImageWithVersion())
 	if err != nil {
 		return "", err
 	}
 
 	extraExecutionFlags := ""
 	if version > 247 {
-		extraExecutionFlags = "--rpc.txfeecap=0"
+		extraExecutionFlags = " --rpc.txfeecap=0"
 	}
 
 	if version > 254 {
 		extraExecutionFlags += " --rpc.allow-unprotected-txs"
 	}
 
+	if version > 255 {
+		extraExecutionFlags += " --db.size.limit=8TB"
+	}
+
 	return extraExecutionFlags, nil
 }
 
-func (g *ErigonPos) buildInitScript() (string, error) {
+func (g *Erigon) buildPosInitScript() (string, error) {
 	extraExecutionFlags, err := g.getExtraExecutionFlags()
 	if err != nil {
 		return "", err
@@ -255,8 +402,7 @@ func (g *ErigonPos) buildInitScript() (string, error) {
 	echo "Starting Erigon..."
 	erigon --http --http.api=eth,erigon,engine,web3,net,debug,trace,txpool,admin --http.addr=0.0.0.0 --http.corsdomain=* \
 		--http.vhosts=* --http.port={{.HttpPort}} --ws --authrpc.vhosts=* --authrpc.addr=0.0.0.0 --authrpc.jwtsecret={{.JwtFileLocation}} \
-		--datadir={{.ExecutionDir}} {{.ExtraExecutionFlags}} --allow-insecure-unlock \
-		--nodiscover --networkid={{.ChainID}} --db.size.limit=8TB`
+		--datadir={{.ExecutionDir}} {{.ExtraExecutionFlags}} --allow-insecure-unlock --nodiscover --networkid={{.ChainID}}`
 
 	data := struct {
 		HttpPort            string
@@ -284,9 +430,45 @@ func (g *ErigonPos) buildInitScript() (string, error) {
 	err = t.Execute(&buf, data)
 
 	return buf.String(), err
+}
+
+func (g *Erigon) buildPowInitScript(minerAddr string) (string, error) {
+	initTemplate := `#!/bin/bash
+	echo "Running erigon init"
+	erigon init /root/genesis.json
+	exit_code=$?
+	if [ $exit_code -ne 0 ]; then
+		echo "Erigon init failed with exit code $exit_code"
+		exit 1
+	fi
+
+	echo "Starting Erigon..."
+	erigon --http --http.api=eth,erigon,engine,web3,net,debug,trace,txpool,admin --http.addr=0.0.0.0 --http.corsdomain=* --http.vhosts=* --http.port={{.HttpPort}} --ws \
+	--allow-insecure-unlock  --nodiscover --networkid={{.ChainID}} --mine --miner.etherbase={{.MinerAddr}} --fakepow`
+
+	data := struct {
+		HttpPort  string
+		ChainID   int
+		MinerAddr string
+	}{
+		HttpPort:  TX_GETH_HTTP_PORT,
+		ChainID:   g.chainConfg.ChainID,
+		MinerAddr: minerAddr,
+	}
+
+	t, err := template.New("init").Parse(initTemplate)
+	if err != nil {
+		fmt.Println("Error parsing template:", err)
+		os.Exit(1)
+	}
+
+	var buf bytes.Buffer
+	err = t.Execute(&buf, data)
+
+	return buf.String(), err
 
 }
 
-func (g *ErigonPos) GetContainerType() ContainerType {
+func (g *Erigon) GetContainerType() ContainerType {
 	return ContainerType_Erigon
 }
