+++ utils/templates/geth.go
@@ -115,22 +115,22 @@ func BuildGenesisJsonForNonDevChain(chainId string, accountAddr []string, extraD
 	return buf.String(), err
 }
 
-type GethGenesisConsensus = string
+type GenesisConsensus = string
 
 const (
 	GethGenesisConsensus_Ethash = "ethash"
 	GethGenesisConsensus_Clique = "clique"
 )
 
-type GethPoWGenesisJsonTemplate struct {
+type GenesisJsonTemplate struct {
 	AccountAddr []string
 	ChainId     string
-	Consensus   GethGenesisConsensus
+	Consensus   GenesisConsensus
 	ExtraData   string
 }
 
 // String representation of the job
-func (c GethPoWGenesisJsonTemplate) String() (string, error) {
+func (c GenesisJsonTemplate) String() (string, error) {
 	var consensusStr string
 	switch c.Consensus {
 	case GethGenesisConsensus_Ethash:
@@ -138,12 +138,13 @@ func (c GethPoWGenesisJsonTemplate) String() (string, error) {
 	case GethGenesisConsensus_Clique:
 		consensusStr = `,"clique": {"period": 1,"epoch": 30000}`
 	default:
-		return "", fmt.Errorf("unsupported consensus type: %s", c.Consensus)
+		// return "", fmt.Errorf("unsupported consensus type: %s", c.Consensus)
+		consensusStr = ""
 	}
 
 	if c.Consensus == GethGenesisConsensus_Clique && (c.ExtraData == "" || c.ExtraData == "0x") {
 		return "", fmt.Errorf("extraData is required for clique consensus")
-	} else if c.Consensus == GethGenesisConsensus_Ethash && c.ExtraData == "" {
+	} else if c.ExtraData == "" {
 		c.ExtraData = "0x"
 	}
